requirements.txt
playwright==1.43.0
python-dotenv
flask
requests
PyPDF2>=3.0.0
google-genai>=0.8.0
google-generativeai